requirements.txt
appdirs==1.4.3
asgiref==3.2.10
beautifulsoup4==4.9.1
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==2.2.10
django-allauth==0.39.1
django-bootstrap-base-template==0.0.2
django-bootstrap4==2.2.0
django-contrib-comments==1.9.2
django-cors-headers==3.0.2
django-countries==6.1.2
django-crispy-forms==1.9.1
django-filter==2.3.0
django-rest-auth==0.9.5
django-summernote==0.8.11.6
django-widget-tweaks==1.4.8
djangorestframework==3.9.4
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
Markdown==3.2.2
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
Pillow==7.0.0
progress==1.5
psycopg2-binary==2.8.5
pyparsing==2.4.6
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.1.0
